backend/requirements.txt
fastapi>=0.129.0
uvicorn[standard]==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
pydantic>=2.9.0
pydantic-settings==2.1.0
email-validator==2.1.0
python-dotenv==1.0.0

# Fix the version conflict:
motor==3.3.2
pymongo==4.6.0

# AI Integration - New google-genai package (replaces deprecated google-generativeai)
google-genai>=1.63.0
